refactor: better types build
