+++ packages/ui-components/vite.config.ts
@@ -2,7 +2,8 @@ import { resolve } from "node:path";
 
 import fs from "fs-extra";
 import { defineConfig } from "vite";
-import dts from "vite-plugin-dts";
+
+// import dts from "vite-plugin-dts";
 
 const packageJson = fs.readJSONSync("package.json");
 
@@ -51,9 +52,9 @@ export default defineConfig({
 		"import.meta.env.BUILDVERSION": JSON.stringify(packageJson.version),
 	},
 	plugins: [
-		dts({
-			include: ["src"],
-			exclude: ["**/__tests__/**/*", "vite-env.d.ts"],
-		}),
+		// dts({
+		// 	include: ["src"],
+		// 	exclude: ["**/__tests__/**/*", "vite-env.d.ts"],
+		// }),
 	],
 });
